+++ src/utils/cdf2coord.c
@@ -26,7 +26,7 @@ int main(int argc,char **argv)
 {
 #if defined(USE_NETCDF) || defined(USE_NETCDF4)
   int rc,ncid,var_id,*dimids,i,j,nvars,lon_id,lat_id,ndims,index,first;
-
+  int index1,index2,len1,len2;
   double *lat,*lon;
   size_t lat_len,lon_len;
   size_t offsets[4]={0,0,0,0},counts[4]={1,1,1,1};
@@ -44,11 +44,11 @@ int main(int argc,char **argv)
   } coord_f;
   char *var;
   FILE *out;
-  Bool lonlat;
+  Bool latlon;
   var=NULL;
   header.datatype=LPJ_SHORT;
   header.scalar=0.01;
-  lonlat=TRUE;
+  latlon=FALSE;
   for(i=1;i<argc;i++)
     if(argv[i][0]=='-')
     {
@@ -73,7 +73,7 @@ int main(int argc,char **argv)
         header.scalar=1;
       }
       else if(!strcmp(argv[i],"-latlon"))
-        lonlat=FALSE;
+        latlon=TRUE;
       else if(!strcmp(argv[i],"-index"))
       {
         if(argc==i+1)
@@ -252,10 +252,23 @@ int main(int argc,char **argv)
   header.cellsize_lat=(float)fabs((lat[lat_len-1]-lat[0])/(lat_len-1));
   fwriteheader(out,&header,LPJGRID_HEADER,LPJGRID_VERSION);
   header.ncell=0;
-  if(lonlat)
-  for(offsets[first]=0;offsets[first]<lat_len;offsets[first]++)
+  if(latlon)
   {
-    for(offsets[first+1]=0;offsets[first+1]<lon_len;offsets[first+1]++)
+    index1=first+1;
+    len1=lon_len;
+    index2=first;
+    len2=lat_len;
+  }
+  else
+  {
+    index1=first;
+    len1=lat_len;
+    index2=first+1;
+    len2=lon_len;
+  }
+  for(offsets[index1]=0;offsets[index1]<len1;offsets[index1]++)
+  {
+    for(offsets[index2]=0;offsets[index2]<len2;offsets[index2]++)
     {
       rc=nc_get_vara_double(ncid,var_id,offsets,counts,&data);
       error(rc);
@@ -286,39 +299,6 @@ int main(int argc,char **argv)
       }
     }
   }
-  else
-    for(offsets[first+1]=0;offsets[first+1]<lon_len;offsets[first+1]++)
-      for(offsets[first]=0;offsets[first]<lat_len;offsets[first]++)
-    {
-      rc=nc_get_vara_double(ncid,var_id,offsets,counts,&data);
-      error(rc);
-      if((!isnan(missing_value) && !isnan(data) && data!=missing_value) ||
-          (isnan(missing_value) && !isnan(data)))
-      {
-        switch(header.datatype)
-        {
-          case LPJ_FLOAT:
-            coord_f.lat=(float)lat[offsets[first]];
-            coord_f.lon=(float)lon[offsets[first+1]];
-            fwrite(&coord_f,sizeof(coord_f),1,out);
-            break;
-          case LPJ_DOUBLE:
-            coord_d.lat=lat[offsets[first]];
-            coord_d.lon=lon[offsets[first+1]];
-            fwrite(&coord_d,sizeof(coord_d),1,out);
-            break;
-          default:
-            coord.lat=(short)(lat[offsets[first]]/header.scalar);
-            coord.lon=(short)(lon[offsets[first+1]]/header.scalar);
-#ifdef DEBUG
-            printf("%.3f %3f %d %d\n",lat[offsets[1]],lon[offsets[2]],coord.lat,coord.lon);
-#endif      
-            fwrite(&coord,sizeof(coord),1,out);
-        }
-        header.ncell++;
-      }
-    }
-
   rewind(out);
   header.firstcell=0;
   header.nyear=1;